Move query string generation logic in generation module
