+++ lib/graphql/language/nodes.rb
@@ -42,43 +42,6 @@ module GraphQL
         def position
           [line, col]
         end
-
-        private
-
-        def generate_value(value)
-          case value
-          when AbstractNode
-            value.to_query_string
-          when FalseClass, Float, Integer, NilClass, String, TrueClass
-            JSON.generate(value, quirks_mode: true)
-          when Array
-            "[#{value.map { |v| generate_value(v) }.join(", ")}]"
-          when Hash
-            "{#{value.map { |k, v| "#{k}: #{generate_value(v)}" }.join(", ")}}"
-          else
-            raise TypeError
-          end
-        end
-
-        def generate_directives(directives)
-          if directives.any?
-            directives.map { |d| " #{d.to_query_string}" }.join
-          else
-            ""
-          end
-        end
-
-        def generate_selections(selections, indent: "")
-          if selections.any?
-            out = " {\n"
-            selections.each do |selection|
-              out << selection.to_query_string(indent: indent + "  ") << "\n"
-            end
-            out << "#{indent}}"
-          else
-            ""
-          end
-        end
       end
 
       class WrapperType < AbstractNode
@@ -115,10 +78,6 @@ module GraphQL
         def children
           [value].flatten.select { |v| v.is_a?(AbstractNode) }
         end
-
-        def to_query_string
-          "#{name}: #{generate_value(value)}"
-        end
       end
 
       class Directive < AbstractNode
@@ -129,10 +88,6 @@ module GraphQL
           @name = name
           @arguments = arguments
         end
-
-        def to_query_string
-          "@#{name}(#{arguments.map(&:to_query_string).join(", ")})"
-        end
       end
 
       class Document < AbstractNode
@@ -142,17 +97,9 @@ module GraphQL
         def initialize_node(definitions: [])
           @definitions = definitions
         end
-
-        def to_query_string
-          definitions.map(&:to_query_string).join("\n")
-        end
       end
 
-      class Enum < NameOnlyNode
-        def to_query_string
-          "#{name}"
-        end
-      end
+      class Enum < NameOnlyNode; end
 
       class Field < AbstractNode
         attr_accessor :name, :alias, :arguments, :directives, :selections
@@ -166,16 +113,6 @@ module GraphQL
           @directives = directives
           @selections = selections
         end
-
-        def to_query_string(indent: "")
-          out = "#{indent}"
-          out << "#{self.alias}: " if self.alias
-          out << "#{name}"
-          out << "(#{arguments.map(&:to_query_string).join(", ")})" if arguments.any?
-          out << generate_directives(directives)
-          out << generate_selections(selections, indent: indent)
-          out
-        end
       end
 
       class FragmentDefinition < AbstractNode
@@ -188,14 +125,6 @@ module GraphQL
           @directives = directives
           @selections = selections
         end
-
-        def to_query_string(indent: "")
-          out = "#{indent}fragment #{name}"
-          out << " on #{type}" if type
-          out << generate_directives(directives)
-          out << generate_selections(selections, indent: indent)
-          out
-        end
       end
 
       class FragmentSpread < AbstractNode
@@ -205,12 +134,6 @@ module GraphQL
         def initialize_node(name: nil, directives: [])
           @name = name
           @directives = directives
-        end
-
-        def to_query_string(indent: "")
-          out = "#{indent}... #{name}"
-          out << generate_directives(directives)
-          out
         end
       end
 
@@ -222,14 +145,6 @@ module GraphQL
           @type = type
           @directives = directives
           @selections = selections
-        end
-
-        def to_query_string(indent: "")
-          out = "#{indent}..."
-          out << " on #{type}" if type
-          out << generate_directives(directives)
-          out << generate_selections(selections, indent: indent)
-          out
         end
       end
 
@@ -248,23 +163,12 @@ module GraphQL
             memo
           end
         end
-
-        def to_query_string
-          generate_value(to_h)
-        end
       end
 
-      class ListType < WrapperType
-        def to_query_string
-          "[#{of_type.to_query_string}]"
-        end
-      end
 
-      class NonNullType < WrapperType
-        def to_query_string
-          "#{of_type.to_query_string}!"
-        end
-      end
+
+      class ListType < WrapperType; end
+      class NonNullType < WrapperType; end
 
       class OperationDefinition < AbstractNode
         attr_accessor :operation_type, :name, :variables, :directives, :selections
@@ -277,22 +181,9 @@ module GraphQL
           @directives = directives
           @selections = selections
         end
-
-        def to_query_string(indent: "")
-          out = "#{indent}#{operation_type}"
-          out << " #{name}" if name
-          out << "(#{variables.map(&:to_query_string).join(", ")})" if variables.any?
-          out << generate_directives(directives)
-          out << generate_selections(selections, indent: indent)
-          out
-        end
       end
 
-      class TypeName < NameOnlyNode
-        def to_query_string
-          "#{name}"
-        end
-      end
+      class TypeName < NameOnlyNode; end
 
       class VariableDefinition < AbstractNode
         attr_accessor :name, :type, :default_value
@@ -301,19 +192,9 @@ module GraphQL
           @type = type
           @default_value = default_value
         end
-
-        def to_query_string
-          out = "$#{name}: #{type.to_query_string}"
-          out << " = #{generate_value(default_value)}" if default_value
-          out
-        end
       end
 
-      class VariableIdentifier < NameOnlyNode
-        def to_query_string
-          "$#{name}"
-        end
-      end
+      class VariableIdentifier < NameOnlyNode; end
     end
   end
 end
